% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bbox_to_sf}
\alias{bbox_to_sf}
\title{Assumes geographic projection
sf bbox to poly}
\usage{
bbox_to_sf(bbox, prj = 4326)
}
\arguments{
\item{bbox}{an sf bbox object}

\item{prj}{defaults to "EPSG:4326"}
}
\description{
Assumes geographic projection
sf bbox to poly
}
\keyword{internal}
