% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bbox_to_sp}
\alias{bbox_to_sp}
\title{Edited from https://github.com/jhollist/quickmapr/blob/master/R/internals.R
sp bbox to poly}
\usage{
bbox_to_sp(bbox)
}
\arguments{
\item{bbx}{an sp object}
}
\description{
Edited from https://github.com/jhollist/quickmapr/blob/master/R/internals.R
sp bbox to poly
}
\keyword{internal}
