% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevatr.R
\name{elevatr}
\alias{elevatr}
\title{Access elevation data from the web}
\description{
This package provides tools to access and download elevation data available
from the Mapzen elevation and Mapzen terrain service.
}

