% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{estimate_raster_size}
\alias{estimate_raster_size}
\title{Estimate download size of DEMs}
\usage{
estimate_raster_size(locations, src, z = NULL)
}
\arguments{
\item{locations}{the locations}

\item{src}{the src}
}
\description{
Estimate download size of DEMs
}
\keyword{internal}
