% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev_profile.R
\name{get_elev_profile}
\alias{get_elev_profile}
\title{Get Point Elevation along a Profile Line}
\usage{
get_elev_profile(
  locations,
  n = NULL,
  density = NULL,
  type = "regular",
  sample = NULL,
  units = NULL,
  include = c("default", "dist", "cumdist"),
  ...,
  prj = NULL,
  overwrite = FALSE,
  coords = c("x", "y"),
  elev_col = "elevation",
  elev_units_col = "elev_units",
  dist_col = "distance"
)
}
\arguments{
\item{locations}{A sf, sfc, or sfg object with LINESTRING geometry.}

\item{n}{integer; number of points to choose per geometry; if missing, n will be computed as \code{round(density * st_length(geom))}.}

\item{density}{numeric; density (points per distance unit) of the sampling, possibly a vector of length equal to the number of features (otherwise recycled); \code{density} may be of class \code{units}.}

\item{type}{character; indicate the sampling type, either "regular" or "random"}

\item{sample}{numeric; a vector of numbers between 0 and 1 indicating the points to sample - if defined sample overrules n, density and type.}

\item{units}{Default: \code{c("meters", "feet")}. Set to "meters" by default. Any
character string that is a valid distance unit supported by
\code{units::set_units()}.}

\item{include}{Option of columns to include: one of "default", "dist", or
"cumdist". Default value returns same columns as \code{\link[=get_elev_point]{get_elev_point()}}. If
\code{include = "dist"}, the returned locations include the distance between
each successive pair of points. If \code{include = "cumdist"}, the distances are
provided as a cumulative sum.}

\item{...}{Additional arguments passed to get_epqs or get_aws_points.  When
using "aws" as the source, pay attention to the \code{z} argument.  A
default of 5 is used, but this uses a raster with a large ~4-5 km
pixel.  Additionally, the source data changes as zoom levels
increase.
Read \url{https://github.com/tilezen/joerd/blob/master/docs/data-sources.md#what-is-the-ground-resolution}
for details.}

\item{prj}{A valid input to \code{\link{st_crs}}.   This
argument is required for a \code{data.frame} of locations and optional
for \code{sf} locations.}

\item{overwrite}{A logical indicating that existing \code{elevation} and
\code{elev_units} columns should be overwritten.  Default is
FALSE and \code{get_elev_point} will error if these columns
already exist.}

\item{coords}{in case of point data: names or numbers of the numeric columns holding coordinates}

\item{elev_col}{Elevation column name.}

\item{elev_units_col}{Elevation units column.}

\item{dist_col}{Column name to use for optional distance column. Only used if
\code{include} is set to \code{"dist"} or \code{"cumdist"}.}
}
\description{
\code{\link[=get_elev_profile]{get_elev_profile()}} allows users to provide LINESTRING inputs to
\code{\link[sf:st_line_sample]{sf::st_line_sample()}} or to cast LINESTRING to POINT before calling
\code{\link[=get_elev_point]{get_elev_point()}} to get the point elevations. The function allows users to
get elevation along a profile line and, optionally, include a distance or
cumulative distance column in the output sf data frame.
}
\examples{
\dontrun{
library(sf)
library(elevatr)

nc <- st_read(system.file("shape/nc.shp", package = "sf")) |>
  st_transform(3857)

nc_line <- suppressWarnings(
  st_cast(
    st_union(
      st_centroid(nc[1, ]),
      st_centroid(nc[2, ])
    ),
    to = "LINESTRING"
  )
)

elev_point <- get_elev_profile(
  nc_line,
  units = "ft",
  dist = TRUE,
  cumulative = TRUE,
  n = 10
)

elev_point

}
}
