% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Get elevation data}
\usage{
get_elevation(location, source = c("epqs", "srtm"), units = c("Meters",
  "Feet"), res = c("30", "90"))
}
\arguments{
\item{location}{A data.frame of the location(s) for which you wish to return 
elevation. The first colum is Longitude and the second 
column is Latitude.}

\item{source}{character string indicating 'epqs' for USGS elevation point 
query service or 'srtm' for Open Topography SRTM service}

\item{units}{Character string of either meters or feet.  Only works for 'epqs'}

\item{res}{Character string of either 30 or 90. Only works for 'srtm' and 
fetches data for the 30 meter SRTM (i.e., SRTM GL1 (Global 30m)) or
the 90 meter SRTM (i.e., SRTM GL3 (Global 90m))}
}
\description{
Primary function for accessing elevation data from a variety of online 
sources
}
\examples{
xdf<-data.frame(runif(100,-80,-72),runif(100,35,45))
get_elevation(xdf)
get_elevation(xdf,"srtm")
}

