% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_epqs}
\alias{get_epqs}
\title{Get EPQS}
\usage{
get_epqs(location, units)
}
\arguments{
\item{location}{location from get_elevation}

\item{units}{units from get_elevation}
}
\description{
Get EPQS
}
\keyword{internal}

