% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev_raster.R
\name{get_gl3}
\alias{get_gl3}
\title{Get a digital elevation model from the Open Topography SRTM Version 3}
\source{
Attribution: Details here
}
\usage{
get_gl3(locations, z, prj, expand = NULL, ...)
}
\arguments{
\item{locations}{Either a \code{data.frame} of x (long) and y (lat), an 
\code{sp}, an \code{sf}, or \code{raster} object as input.}

\item{prj}{PROJ.4 string for input bbox}

\item{expand}{A numeric value of a distance, in map units, used to expand the
bounding box that is used to fetch the SRTM data.}

\item{...}{Extra configuration parameters to be passed to httr::GET.  Common 
usage is to adjust timeout.  This is done as 
\code{config=timeout(x)} where \code{x} is a numeric value in 
seconds.  Multiple configuration functions may be passed as a 
vector.}
}
\description{
This function uses the Open Topography SRTM Version 3 files.
}
\keyword{internal}
