% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_mapzen_terrain}
\alias{get_mapzen_terrain}
\title{Get a digital elevation model from the Mapzen Terrain Tiles}
\usage{
get_mapzen_terrain(bbx, z = 10, api_key = getOption("mapzen_key"),
  file_format = c("geotiff"))
}
\arguments{
\item{bbx}{a \code{sp::bbox} object that defines the area to return}

\item{z}{The zoom level to return.  The zoom ranges from 1 to 15.  Resolution
of the resultant raster is determined by the zoom and latitude.  For 
details on zoom and resolution see the documentation from Mapzen at 
\url{https://mapzen.com/documentation/terrain-tiles/data-sources/#what-is-the-ground-resolution}}

\item{api_key}{An API Key from Mapzen, create at 
\url{https://mapzen.com/developer} Required. Set in your 
\code{.Rprofile} file with the option \code{mapzen_key}}

\item{file_format}{Several file formats are provided by Mapzen: terrarium, 
normal, geotiff, and skadi.  Currently only the geotiff
format is supported}
}
\description{
This function uses the Mapzen Terrain Tile service to retrieve an elevation
raster.  It accepts a \code{sp::bbox} object as input and returns a single
raster object covering that extent.  You must have an api_key from Mapzen.
}
\examples{
library(quickmapr)
library(sp)
data(lake)
wgs84_dd <- "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"
lake_dd <- spTransform(lake,CRS(wgs84_dd))
get_mapzen_terrain(bbox(lake_dd),z=13)
}

