% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_opentopo_key}
\alias{get_opentopo_key}
\title{OpenTopo Key}
\usage{
get_opentopo_key()
}
\description{
The OpenTopography API now requires an API Key.  This function will grab your
key from an .Renviron file
}
\keyword{internal}
