% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_srtm}
\alias{get_srtm}
\title{Get SRTM from Open Topography}
\usage{
get_srtm(location, res)
}
\arguments{
\item{location}{location from get_elevation}

\item{res}{res from get_elevation}
}
\description{
Get SRTM from Open Topography
}
\keyword{internal}

