% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_tilexy}
\alias{get_tilexy}
\title{function to get a data.frame of all xyz tiles to download}
\usage{
get_tilexy(bbx, z)
}
\description{
function to get a data.frame of all xyz tiles to download
}
\keyword{internal}

