% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lake}
\alias{lake}
\title{SpatialPolygonsDataFrame of Lake Sunapee}
\format{
SpatialPolygonDataframe with 1 lakes, each with 13 variables
}
\description{
This example data is a SpatialPolygonsDataFrame 
of a single lake, Lake Sunapee.  Used for examples and tests.
}
\keyword{datasets}
