% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{latlong_to_tilexy}
\alias{latlong_to_tilexy}
\title{function to convert lat long to xyz tile with decimals
rounding to tile occurs in \code{get_tilexy}}
\usage{
latlong_to_tilexy(lon_deg, lat_deg, zoom)
}
\description{
function to convert lat long to xyz tile with decimals
rounding to tile occurs in \code{get_tilexy}
}
\keyword{internal}
