% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{loc_check}
\alias{loc_check}
\title{function to get a data.frame of all xyz tiles to download
function to check input type and projection.  All input types convert to a
SpatialPointsDataFrame for point elevation and bbx for raster.}
\usage{
loc_check(locations, prj = NULL)
}
\description{
function to get a data.frame of all xyz tiles to download
function to check input type and projection.  All input types convert to a
SpatialPointsDataFrame for point elevation and bbx for raster.
}
\keyword{internal}
