% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{loc_linestring_to_point}
\alias{loc_linestring_to_point}
\title{function sf, sfc, or sfg with LINESTRING geometry to POINT geometry}
\usage{
loc_linestring_to_point(
  locations,
  n = NULL,
  density = NULL,
  type = "regular",
  sample = NULL
)
}
\arguments{
\item{locations}{A sf, sfc, or sfg object with LINESTRING geometry.}

\item{n}{integer; number of points to choose per geometry; if missing, n will be computed as \code{round(density * st_length(geom))}.}

\item{density}{numeric; density (points per distance unit) of the sampling, possibly a vector of length equal to the number of features (otherwise recycled); \code{density} may be of class \code{units}.}

\item{type}{character; indicate the sampling type, either "regular" or "random"}

\item{sample}{numeric; a vector of numbers between 0 and 1 indicating the points to sample - if defined sample overrules n, density and type.}
}
\description{
Optionally use \code{\link[sf:st_line_sample]{sf::st_line_sample()}} when n, density, or sample are
supplied.
}
\keyword{internal}
