% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{mapzen_elev_GET_nokey}
\alias{mapzen_elev_GET_nokey}
\title{based on https://mapzen.com/documentation/overview/#rate-limits
limits for keyless access are 1/second and 6/minute
see issue #4}
\usage{
mapzen_elev_GET_nokey(url = NULL, config = list(), ..., handle = NULL)
}
\description{
based on https://mapzen.com/documentation/overview/#rate-limits
limits for keyless access are 1/second and 6/minute
see issue #4
}
\keyword{internal}

