% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{mapzen_elev_GET_withkey}
\alias{mapzen_elev_GET_withkey}
\title{based on https://mapzen.com/documentation/overview/#rate-limits
limits for valid keyholders are 2/second and 20K/day
this function will only enforce the 2/second limit
see issue #4}
\usage{
mapzen_elev_GET_withkey(url = NULL, config = list(), ..., handle = NULL)
}
\description{
based on https://mapzen.com/documentation/overview/#rate-limits
limits for valid keyholders are 2/second and 20K/day
this function will only enforce the 2/second limit
see issue #4
}
\keyword{internal}
