% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev_raster.R
\name{merge_rasters}
\alias{merge_rasters}
\title{Merge Rasters}
\usage{
merge_rasters(
  raster_list,
  target_prj,
  method = "bilinear",
  returnRaster = TRUE
)
}
\arguments{
\item{raster_list}{a list of raster file paths to be mosaiced}

\item{target_prj}{the target projection of the output raster}

\item{method}{the method for resampling/reprojecting. Default is 'bilinear'. 
Options can be found [here](https://gdal.org/programs/gdalwarp.html#cmdoption-gdalwarp-r)}

\item{returnRaster}{if TRUE, return a raster object (default), else, return the file path to the object}
}
\description{
Merge multiple downloaded raster files into a single file. The input `target_prj` 
describes the projection for the new grid.
}
\keyword{internal}
