% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{proj_expand}
\alias{proj_expand}
\title{function to project bounding box and if needed expand it}
\usage{
proj_expand(bbx, prj, expand)
}
\description{
function to project bounding box and if needed expand it
}
\keyword{internal}
