% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_df}
\alias{pt_df}
\title{Small data frame of xy locations}
\format{A data.frame with two columns, x(long) and y(lat)}
\description{
Example data frame of locations for use in examples and text
}
\keyword{datasets}
