% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{relocate_sf_col_end}
\alias{relocate_sf_col_end}
\title{Make sf column the last column in a sf data frame}
\usage{
relocate_sf_col_end(x)
}
\description{
Make sf column the last column in a sf data frame
}
\keyword{internal}
