% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{set_api_key}
\alias{set_api_key}
\title{Utility function to set API key from mapzen}
\usage{
set_api_key(api_key, src = c("mapzen"))
}
\description{
This function updates the .Renviron file with a new key
}
\keyword{internal}
