% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_big}
\alias{sf_big}
\title{A sf POINT dataset of random points}
\format{
A sf POINT object
}
\description{
This sf POINT dataset is 250 uniform random points to be used for 
examples and tests
}
\keyword{datasets}
