% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_big}
\alias{sp_big}
\title{SpatialPoints data}
\format{A SpatialPoints object}
\description{
This SpatialPoints dataset is 250 uniform random points to be used for 
examples and tests
}
\keyword{datasets}

