% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{st_point_distances}
\alias{st_point_distances}
\title{Get distances between successive pairs of points}
\usage{
st_point_distances(
  x,
  cumulative = TRUE,
  units = NULL,
  prj = sf::st_crs(x),
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[sf:geos_measures]{sf::st_distance}}
  \describe{
    \item{\code{dist_fun}}{deprecated}
    \item{\code{by_element}}{logical; if \code{TRUE}, return a vector with distance between the first elements of \code{x} and \code{y}, the second, etc; an error is raised if \code{x} and \code{y} are not the same length. If \code{FALSE}, return the dense matrix with all pairwise distances.}
    \item{\code{which}}{character; for Cartesian coordinates only: one of \code{Euclidean}, \code{Hausdorff} or \code{Frechet}; for geodetic coordinates, great circle distances are computed; see details}
    \item{\code{par}}{for \code{which} equal to \code{Hausdorff} or \code{Frechet}, optionally use a value between 0 and 1 to densify the geometry}
    \item{\code{tolerance}}{ignored if \code{st_is_longlat(x)} is \code{FALSE}; otherwise, if set to a positive value, the first distance smaller than \code{tolerance} will be returned, and true distance may be smaller; this may speed up computation. In meters, or a \code{units} object convertible to meters.}
  }}
}
\description{
Get distances between successive pairs of points
}
\keyword{internal}
