% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev_point.R
\name{get_aws_points}
\alias{get_aws_points}
\title{Get point elevation data from the AWS Terrain Tiles}
\usage{
get_aws_points(locations, units = c("meters", "feet"), ...)
}
\arguments{
\item{locations}{Either a \code{data.frame} with x (e.g. longitude) as the 
first column and y (e.g. latitude) as the second column, a 
\code{SpatialPoints}/\code{SpatialPointsDataFrame}, or a 
\code{sf} \code{POINT} or \code{MULTIPOINT} object.   
Elevation for these points will be returned in the 
originally supplied class.}

\item{units}{Character string of either meters or feet. Conversions for 
'aws' are handled in R as the AWS terrain tiles are served in 
meters.}

\item{...}{Arguments to be passed to \code{get_elev_raster}}

\item{z}{The zoom level to return.  The zoom ranges from 1 to 14.  Resolution
of the resultant raster is determined by the zoom and latitude.  For 
details on zoom and resolution see the documentation from Mapzen at 
\url{https://mapzen.com/documentation/terrain-tiles/data-sources/#what-is-the-ground-resolution}.  
Returned}
}
\value{
a SpatialPointsDataFrame or sf POINT or MULTIPOINT object with 
        elevation added to the data slot
}
\description{
Function for accessing elevation data from AWS and extracting the elevations
}
\examples{
data(lake)
lake_sf <- sf::st_as_sf(lake)
loc_df <- data.frame(x = runif(6,min=sp::bbox(lake)[1,1], 
                               max=sp::bbox(lake)[1,2]),
                     y = runif(6,min=sp::bbox(lake)[2,1], 
                               max=sp::bbox(lake)[2,2]))
loc_sf <- sf::st_as_sf(loc_df, coords = c("x", "y"), crs = sf::st_crs(lake_sf))                               
get_elev_point(loc_sf, src = "aws", z = 12)
}
\keyword{internal}
