% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev_raster.R
\name{get_elev_raster}
\alias{get_elev_raster}
\title{Get Raster Elevation}
\usage{
get_elev_raster(locations, prj, source = c("mapzen"), api_key = NULL, ...)
}
\arguments{
\item{locations}{Either a \code{data.frame} of x (long) and y (lat), an 
\code{sp}, or \code{raster} object as input.}

\item{prj}{A PROJ.4 string defining the projection of the locations argument. 
If a \code{sp} or \code{raster} object is provided, the PROJ.4 string 
will be taken from that.  This argument is required for a 
\code{data.frame} of locations.}

\item{source}{A character indicating which API to use, currently only 
"mapzen" is used.}

\item{api_key}{A valid API key.}

\item{...}{Extra parameters to pass to API specific fucntions}
}
\value{
Function returns a \code{SpatialPointsDataFrame} in the projection 
        specified by the \code{prj} argument.
}
\description{
Several web services provide access to raster elevation. Currently, this 
function provides access to the Mapzen Terrain Service The function 
accepts a \code{data.frame} of x (long) and y (lat), an 
\code{sp}, or \code{raster} object as input.  A \code{raster} object is returned.
}
\examples{
data(lake)
loc_df <- data.frame(x = runif(6,min=bbox(lake)[1,1], max=bbox(lake)[1,2]),
                     y = runif(6,min=bbox(lake)[2,1], max=bbox(lake)[2,2]))
get_elev_raster(locations = loc_df, prj = proj4string(lake))
}

