% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev_point.R
\name{get_mapzen_elev}
\alias{get_mapzen_elev}
\title{Get point elevations from Mapzen}
\source{
Attribution: Mapzen terrain tiles contain 3DEP, SRTM, and GMTED2010 
        content courtesy of the U.S. Geological Survey and ETOPO1 content 
        courtesy of U.S. National Oceanic and Atmospheric Administration. 
        \url{https://mapzen.com/documentation/elevation/elevation-service/}
}
\usage{
get_mapzen_elev(locations, api_key = getOption("mapzen_key"))
}
\arguments{
\item{api_key}{A valid Mapzen API key.  Not required, but higher rate limits
are allowed with a key. Defaults to 
\code{getOption("mapzen_key")}.}
}
\description{
Get point elevations from Mapzen
}
\keyword{internal}

