% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev_point.R
\name{get_mapzen_elev}
\alias{get_mapzen_elev}
\title{Get point elevations from Mapzen}
\source{
Attribution: Mapzen terrain tiles contain 3DEP, SRTM, and GMTED2010 
        content courtesy of the U.S. Geological Survey and ETOPO1 content 
        courtesy of U.S. National Oceanic and Atmospheric Administration. 
        \url{https://mapzen.com/documentation/elevation/elevation-service/}
}
\usage{
get_mapzen_elev(locations, api_key = Sys.getenv("mapzen_key"))
}
\arguments{
\item{api_key}{A valid Mapzen API key.  Although not required by the API, the
rate limits are low without a key.  The \code{elevatr} package
requires a key.  To get a key, visit 
\url{https://mapzen.com/developers}. Defaults to 
\code{Sys.getenv("mapzen_key")}.}
}
\description{
Get point elevations from Mapzen
}
\keyword{internal}

