% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{loc_check}
\alias{loc_check}
\title{function to check and prepare input locations}
\usage{
loc_check(locations, prj = NULL, coords = c("x", "y"), elev_col = "elevation")
}
\arguments{
\item{coords}{in case of point data: names or numbers of the numeric columns holding coordinates}

\item{elev_col}{Elevation column name.}
}
\description{
All input types convert to a sf data frame.
}
\keyword{internal}
