% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{loc_check}
\alias{loc_check}
\title{function to check input type and projection and convert data frame to 
SpatialPoints}
\usage{
loc_check(locations, prj = NULL)
}
\description{
function to check input type and projection and convert data frame to 
SpatialPoints
}

