% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{set_api_key}
\alias{set_api_key}
\title{function to set api key in .Renviron}
\usage{
set_api_key(key, name)
}
\arguments{
\item{key}{the api key}

\item{name}{the name to assign to the key in your .Renviron}
}
\description{
function to set api key in .Renviron
}
